added js modules
